共通部分がない独立した処理をクラスを分けた
